rename nnf->nnfusion, change mailbox
